pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nv-c13-tomography-qml"
version = "0.1.0"
description = "NV 13C tomography + QML utilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "matplotlib",
  "scipy",
  "pyyaml",
  "numba"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
